+++ Arma3ObjectBuilder/ui/__init__.py
@@ -3,7 +3,6 @@ from . import import_export_mcfg
 from . import import_export_p3d
 from . import import_export_rtm
 from . import import_export_tbcsv
-from . import import_export_paa
 from . import props_action
 from . import props_material
 from . import props_object_mesh
